% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_changepoint_plot_helpers.R
\name{add_changepoints}
\alias{add_changepoints}
\title{Adds changepoints annotations to a plotly plot.}
\usage{
add_changepoints(plotly_obj, original_data, changepoint_data)
}
\arguments{
\item{plotly_obj}{a plotly object}

\item{original_data}{A dataset with two columns}

\item{changepoint_data}{The output of gfpop::gfpop() with original_data}
}
\value{
A list of 3 objects: "plot" (the plotly plot with changepoints added),
"changepoint_annotations", and "changepoint_annotations_regions"
}
\description{
Adds changepoints annotations to a plotly plot.
}
