% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_visNetwork_helpers.R
\name{add_edge}
\alias{add_edge}
\title{Add a new edge to a visNetwork edge dataframe}
\usage{
add_edge(
  edgedf,
  id,
  label,
  to,
  from,
  type,
  parameter,
  penalty,
  K = Inf,
  a = 0,
  min = NA,
  max = NA,
  selfReference.angle = NA,
  selfReference.size = 40,
  hidden = FALSE,
  color = "black"
)
}
\arguments{
\item{edgedf}{a dataframe containing the edges.}

\item{id}{edge id}

\item{label}{edge label}

\item{to}{Where does the edge go to?}

\item{from}{Where does the edge come from?}

\item{type}{What type of edge?}

\item{parameter}{parameter: gap, etc.}

\item{penalty}{edge penalty}

\item{K}{K (see gfpop)}

\item{a}{a (see gfpop)}

\item{min}{min (see gfpop)}

\item{max}{max (see gfpop)}

\item{selfReference.angle}{The angle of this edge, if it's recursive}

\item{selfReference.size}{The length of this edge, if it's recursive}

\item{hidden}{Is this edge hidden?}

\item{color}{edge color}
}
\value{
a dataframe with one more row than edgedf
}
\description{
Add a new edge to a visNetwork edge dataframe
}
