% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_visNetwork_helpers.R
\name{add_null_edge}
\alias{add_null_edge}
\title{Adds a recursive null edge}
\usage{
add_null_edge(edgedf, nodeid)
}
\arguments{
\item{edgedf}{a dataframe containing the edges}

\item{nodeid}{the id of the node on which to create this recursive edge}
}
\value{
a dataframe with one more row than edgedf
}
\description{
Adds a recursive null edge
}
