% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_changepoint_helpers.R
\name{annotate_data_with_changepoint}
\alias{annotate_data_with_changepoint}
\title{Given data and it's associated changepoint data, annotates data points
with the changepoint interval in which that data falls}
\usage{
annotate_data_with_changepoint(data_input, changepoint_data)
}
\arguments{
\item{data_input}{The data associated with the given changepoints}

\item{changepoint_data}{The changepoints, in a dataframe format 
(see generate_changepoint)}
}
\value{
A dataframe of the data_input annotated with associated changepoint
}
\description{
Given data and it's associated changepoint data, annotates data points
with the changepoint interval in which that data falls
}
\examples{
data <- data.frame(X = 1:10, Y = gfpop::dataGenerator(10, c(0.1, 0.3, 0.5, 0.8, 1), 
c(1, 2, 1, 3, 1), sigma = 1))
graph <- gfpop::graph(type = "std")
changepoint_data <- generate_changepoint(data$Y, graph)
annotate_data_with_changepoint(data, changepoint_data)
}
