% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_visNetwork_helpers.R
\name{create_label}
\alias{create_label}
\title{Creates a 'label' array for a graphdf, given the columns that you
want to include in the label and the seperator}
\usage{
create_label(graphdf, columns = c("type", "penalty"), collapse = " | ")
}
\arguments{
\item{graphdf}{A graph object (in the form of a data.table) from gfpop
or the "edges" from a visNetwork list.}

\item{columns}{a character vector of columns to be included in the label.}

\item{collapse}{A string to seperate each item}
}
\value{
a character vector combining those columns as specified
}
\description{
Creates a 'label' array for a graphdf, given the columns that you
want to include in the label and the seperator
}
