% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_visNetwork_helpers.R
\name{create_label_individual}
\alias{create_label_individual}
\title{Creates an individual label for an individual edge}
\usage{
create_label_individual(
  state1,
  state2,
  type,
  parameter,
  penalty,
  K,
  a,
  min,
  max,
  columns = c("type", "penalty"),
  collapse = " | "
)
}
\arguments{
\item{state1}{The state where the edge originates}

\item{state2}{The state to which the edge travels}

\item{type}{The type of edge (see gfpop::graph types)}

\item{parameter}{A parameter, such as '1' for null, or gap otherwise}

\item{penalty}{An edge penalty (see gfpop::graph)}

\item{K}{see gfpop::graph}

\item{a}{see gfpop::graph}

\item{min}{see gfpop::graph}

\item{max}{see gfpop::graph}

\item{columns}{The columns to include in this edge}

\item{collapse}{the character separating edge columns in this label}
}
\description{
Creates an individual label for an individual edge
}
