% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{details}
\alias{details}
\title{Create a details drop-down with the given summary and content}
\usage{
details(summary, content, summary_multiplier = 1)
}
\arguments{
\item{summary}{Content for the 'summary' component of the details tag}

\item{content}{Content to be put in the details tag}

\item{summary_multiplier}{A size multiplier for the summary size. 2=2X, etc.}
}
\value{
an HTML object
}
\description{
Create a details drop-down with the given summary and content
}
