% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_graph_helpers.R
\name{format_edge}
\alias{format_edge}
\title{Takes in a row from a graph dataframe, returns that row formatted as R code}
\usage{
format_edge(edge_df)
}
\arguments{
\item{edge_df}{One row from a gfpop::graph() graph df, with column names}
}
\value{
a string corresponding to the code that, when run, produces the 
given edge
}
\description{
Takes in a row from a graph dataframe, returns that row formatted as R code
}
\examples{
graph <- gfpop::graph(type = "std")
format_edge(graph[1,])
}
