% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_changepoint_helpers.R
\name{generate_changepoint}
\alias{generate_changepoint}
\title{Given data and a constraint graph, generate a changepoint output
A wrapper around gfpop::gfpop()}
\usage{
generate_changepoint(data_input, graph_input, type = "mean")
}
\arguments{
\item{data_input}{vector of data to segment}

\item{graph_input}{dataframe of class "graph" to constraint the changepoint inference}

\item{type}{a string defining the cost model to use: "mean", "variance", "poisson",
"exp", "negbin"}
}
\value{
A dataframe with columns:
\itemize{
\item{changepoint}{a vector of X locations where the changepoints begin}
\item{changepoint_end}{a vector of X locations where the changepoints end}
\item{y}{A vector of Y locations where the changepoints are centered}
}
}
\description{
Given data and a constraint graph, generate a changepoint output
A wrapper around gfpop::gfpop()
}
\examples{
data <- gfpop::dataGenerator(10, c(0.1, 0.3, 0.5, 0.8, 1), 
c(1, 2, 1, 3, 1), sigma = 1)
graph <- gfpop::graph(type = "std")
generate_changepoint(data, graph)
}
