% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_visNetwork_helpers.R
\name{generate_visNetwork}
\alias{generate_visNetwork}
\title{Generates a visNetwork from a list of nodes and edges}
\usage{
generate_visNetwork(graph_data)
}
\arguments{
\item{graph_data}{A list containing 'nodes' and 'edges' is visNetwork format}
}
\value{
a visNetwork object
}
\description{
Generates a visNetwork from a list of nodes and edges
}
\examples{
generate_visNetwork(graphdf_to_visNetwork(gfpop::graph(type = "std")))
}
