% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_changepoint_helpers.R
\name{get_associated_changepoint}
\alias{get_associated_changepoint}
\title{Gets the changepoint location associated with this location on the x axis}
\usage{
get_associated_changepoint(x_loc, changepoint_data)
}
\arguments{
\item{x_loc}{The X locations to querey}

\item{changepoint_data}{A dataframe containing changepoint data 
(see generate_changepoint).}
}
\value{
A data.frame/list with the 0 or 1 changepoint location consistent
with the given x location
}
\description{
Gets the changepoint location associated with this location on the x axis
}
\examples{
data <- gfpop::dataGenerator(10, c(0.1, 0.3, 0.5, 0.8, 1), 
c(1, 2, 1, 3, 1), sigma = 1)
graph <- gfpop::graph(type = "std")
changepoint_data <- generate_changepoint(data, graph)
get_associated_changepoint(3, changepoint_data)
}
