% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_graph_helpers.R
\name{graph_to_R_code}
\alias{graph_to_R_code}
\title{Takes in a graph dataframe, returns the graph formatted as R code}
\usage{
graph_to_R_code(graph)
}
\arguments{
\item{graph}{A graph df, like that returned by gfpop::graph()}
}
\value{
a string corresponding to the code that, when run, produces the
given graph
}
\description{
Takes in a graph dataframe, returns the graph formatted as R code
}
\examples{
graph <- gfpop::graph(type = "std")
graph_to_R_code(graph)
}
