% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_visNetwork_helpers.R
\name{graphdf_to_visNetwork}
\alias{graphdf_to_visNetwork}
\title{Turns a graph dataframe (from gfpop) into a list that
can be read by visNetwork}
\usage{
graphdf_to_visNetwork(graphdf, edgeSep = "_")
}
\arguments{
\item{graphdf}{A graph object (in the form of a dataframe) from gfpop
Must have the following columns:
\itemize{
\item{"state1"}
\item{"state2"}
\item{"}
}}
}
\value{
a list that can be read by visNetwork
}
\description{
Turns a graph dataframe (from gfpop) into a list that
can be read by visNetwork
}
\examples{
graphdf_to_visNetwork(gfpop::graph(type = "std"))
}
