% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{inline_div}
\alias{inline_div}
\title{Creates an inline block div component aligned top with the given width percentage}
\usage{
inline_div(width, ...)
}
\arguments{
\item{width}{numeric between 0 and 1, width of the div component}

\item{...}{Parameters to be put inside the div, usually some shiny UI elem}
}
\value{
a UI HTML div
}
\description{
Creates an inline block div component aligned top with the given width percentage
}
