% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_changepoint_plot_helpers.R
\name{make_changepoint}
\alias{make_changepoint}
\title{Make a changepoint data.table}
\usage{
make_changepoint(
  parameter_x,
  region_y,
  new_state,
  previous_state,
  changepoint_number = NaN,
  add_na = TRUE
)
}
\arguments{
\item{parameter_x}{The x location of this changepoint}

\item{region_y}{A range of values between the mean of the previous changeregion
and the mean of the next changeregion}

\item{new_state}{The name of the next changeregion's state}

\item{previous_state}{The name of the previous changeregion's state}

\item{changepoint_number}{The index/number of this changepoint (default NaN)}

\item{add_na}{Whether to add a row of NAs at the bottom of the return table}
}
\value{
(data.table) describing the changepoint
}
\description{
Make a changepoint data.table
}
