% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{make_changepoint_locs}
\alias{make_changepoint_locs}
\title{Make a vector of relative changepoint locations to use as input to gfpop's
generateData function.}
\usage{
make_changepoint_locs(n, e)
}
\arguments{
\item{n}{int, the number of changepoints}

\item{e}{numeric, the logarithmic spacing between changepoints}
}
\description{
Make a vector of relative changepoint locations to use as input to gfpop's
generateData function.
}
