% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_changepoint_plot_helpers.R
\name{make_region}
\alias{make_region}
\title{Make a changeregion data.table}
\usage{
make_region(region_x, parameter_y, state, next_cp = NaN, add_na = TRUE)
}
\arguments{
\item{region_x}{The x values over which this region spans}

\item{parameter_y}{The median of this region}

\item{state}{The state of this region}

\item{next_cp}{The median of the next region}

\item{add_na}{Whether to add a row of NAs at the bottom of the return table}
}
\value{
(data.table) describing the changeregion
}
\description{
Make a changeregion data.table
}
