% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_analysis.R
\name{mod_analysis_server}
\alias{mod_analysis_server}
\title{analysis Server Function}
\usage{
mod_analysis_server(id, gfpop_data = reactiveValues())
}
\arguments{
\item{id}{The id of this module}

\item{gfpop_data}{The main data for the gfpop analysis, passed from home mod}
}
\description{
analysis Server Function
}
