% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_home.R
\name{mod_home_server_core}
\alias{mod_home_server_core}
\title{Core functionality of the home server function}
\usage{
mod_home_server_core(input, output, session)
}
\arguments{
\item{input}{From shiny}

\item{output}{From shiny}

\item{session}{From shiny}
}
\description{
Core functionality of the home server function
}
