% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_visNetwork_helpers.R
\name{modify_visNetwork}
\alias{modify_visNetwork}
\title{Edits a visNetwork data based on the given command, consistent with the format
of commands passed to the _graphChange observers in visNetwork-shiny}
\usage{
modify_visNetwork(event, graphdata_visNetwork)
}
\arguments{
\item{event}{The event, with at least a $cmd entry, based on observer format}

\item{graphdata_visNetwork}{A list of data to be passed to visNetwork}
}
\value{
the same format as graphdata_visNetwork, but edited according to event
}
\description{
Edits a visNetwork data based on the given command, consistent with the format
of commands passed to the _graphChange observers in visNetwork-shiny
}
