% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{mutate_cond}
\alias{mutate_cond}
\title{Like dplyr::mutate, but mutations on a condition **by row**, not by column.}
\usage{
mutate_cond(.data, condition, ..., envir = parent.frame())
}
\arguments{
\item{.data}{User data}

\item{condition}{Condition to select by}

\item{...}{Args mased to dplyr::mutate}

\item{envir}{(default: parent.frame())}
}
\description{
Like dplyr::mutate, but mutations on a condition **by row**, not by column.
}
