% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_visNetwork_helpers.R
\name{NAtoNone}
\alias{NAtoNone}
\title{Converts all NA objects in a vector to "None"}
\usage{
NAtoNone(vec)
}
\arguments{
\item{vec}{a vector or unnamed list (vector equivalent)}
}
\value{
a unnamed list (vector equivalent) where "None"-->NA
}
\description{
Converts all NA objects in a vector to "None"
}
