% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_visNetwork_helpers.R
\name{NonetoNA}
\alias{NonetoNA}
\title{Converts all "None" objects in a vector into NA}
\usage{
NonetoNA(vec)
}
\arguments{
\item{vec}{a vector or unnamed list (vector equivalent)}
}
\value{
a unnamed list (vector equivalent) where "None"-->NA
}
\description{
Converts all "None" objects in a vector into NA
}
