% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_graph_helpers.R
\name{select_graph_columns}
\alias{select_graph_columns}
\title{Selects the columns from this graph dataframe that are relevant to gfpop}
\usage{
select_graph_columns(graph_df)
}
\arguments{
\item{graph_df}{the graph to process}
}
\value{
a dataframe with 9 columns, compatable with gfpop
}
\description{
Selects the columns from this graph dataframe that are relevant to gfpop
}
\examples{
graph <- gfpop::graph(type = "std")
graph$test <- c(1,2)
select_graph_columns(graph)
}
