% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_visNetwork_helpers.R
\name{update_visNetwork_edges}
\alias{update_visNetwork_edges}
\title{Updates the given visNetwork edges, based on a new graphdf,
excluding updates to certain columns. This is useful as to not mess up
visNetwork edge IDs, which are important to keeping visNetwork happy.}
\usage{
update_visNetwork_edges(
  old_edges,
  graphdf,
  edgeSep = "_",
  showNull = TRUE,
  label_columns = c("type", "penalty"),
  columns_to_exclude = c("id", "to", "from")
)
}
\arguments{
\item{old_edges}{A dataframe in visNetwork edges format}

\item{graphdf}{A graph object (in the form of a dataframe) from gfpop}

\item{edgeSep}{A character seperating the nodes in an edge label}

\item{label_columns}{(character) An array of columns to use to make edge labels}

\item{columns_to_exclude}{(character) An array of columns not to include in the update}

\item{hideNull}{(Boolean) hide null edges?}
}
\description{
Updates the given visNetwork edges, based on a new graphdf,
excluding updates to certain columns. This is useful as to not mess up
visNetwork edge IDs, which are important to keeping visNetwork happy.
}
