% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_visNetwork_helpers.R
\name{visNetwork_to_graphdf}
\alias{visNetwork_to_graphdf}
\title{Turns a list that can be read by visNetwork into graph dataframe (for gfpop)}
\usage{
visNetwork_to_graphdf(visNetwork_list)
}
\arguments{
\item{visNetwork_list}{A list object compatable with visNetwork
See graphdf_to_visNetwork.}
}
\value{
a dataframe/graph for gfpop
}
\description{
Turns a list that can be read by visNetwork into graph dataframe (for gfpop)
}
\examples{
visNetwork_to_graphdf(graphdf_to_visNetwork(gfpop::graph(type = "std")))
}
