% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_visNetwork_helpers.R
\name{add_node}
\alias{add_node}
\title{Add a new node to a visNetwork node dataframe}
\usage{
add_node(
  nodedf,
  id,
  label = "",
  size = 40,
  start = FALSE,
  end = FALSE,
  shape = "dot",
  color.background = "lightblue",
  color.border = "lightblue",
  shadow = FALSE
)
}
\arguments{
\item{nodedf}{a dataframe containing the nodes.}

\item{id}{id of the new node.}

\item{label}{label of the new node. default: ""}

\item{size}{size of the new node. Default: 40}

\item{start}{is this node a start node? Default: FALSE}

\item{end}{is this node an end node? Default: FALSE}

\item{shape}{the shape of this node. See visNetwork docs. Default: "dot"}

\item{color.background}{the background color of the node. Default: "lightblue"}

\item{color.border}{the border color of the node. Default: "lightblue"}

\item{shadow}{Whether this node has a shadow. Default: false}
}
\value{
a dataframe with one more row than nodedf
}
\description{
Add a new node to a visNetwork node dataframe
}
