% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_visNetwork_helpers.R
\name{create_label_individual}
\alias{create_label_individual}
\title{Creates an individual label for an individual edge}
\usage{
create_label_individual(
  state1,
  state2,
  type,
  parameter,
  penalty,
  K,
  a,
  min,
  max,
  columns = c("type", "penalty"),
  collapse = " | "
)
}
\description{
Creates an individual label for an individual edge
}
