% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_visNetwork_helpers.R
\name{graphdf_to_visNetwork}
\alias{graphdf_to_visNetwork}
\title{Turns a graph dataframe (from gfpop) into a list that
can be read by visNetwork}
\usage{
graphdf_to_visNetwork(
  graphdf,
  edgeSep = "_",
  showNull = TRUE,
  label_columns = c("type", "penalty"),
  label_collapse = " | ",
  edge_ids = c()
)
}
\arguments{
\item{graphdf}{A graph object (in the form of a dataframe) from gfpop
Must have the following columns:
\itemize{
\item{"state1"}
\item{"state2"}
\item{"type"}
\item{"parameter"}
\item{"penalty"}
\item{"K"}
\item{"a"}
\item{"min"}
\item{"max"}
}}

\item{edgeSep}{A character seperating the nodes in an edge label}

\item{showNull}{(Boolean) hide null edges?}

\item{label_columns}{(character) An array of columns to use to make edge labels}

\item{label_collapse}{the character separating edge columns in this label}

\item{edge_ids}{An array of ids, one for each edge in the graphdf}
}
\value{
a list that can be read by visNetwork
}
\description{
Turns a graph dataframe (from gfpop) into a list that
can be read by visNetwork
}
\examples{
graphdf_to_visNetwork(gfpop::graph(type = "std"), edge_ids = 1:2)
}
