% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_changepoint_plot_helpers.R
\name{plot_changepoint}
\alias{plot_changepoint}
\title{Makes a ggplot object with the changepoint data}
\usage{
plot_changepoint(data_input, changepoint_data)
}
\arguments{
\item{data_input}{The data associated with the given changepoints}

\item{changepoint_data}{The changepoints, in a dataframe format}
}
\value{
a ggplot object
}
\description{
Makes a ggplot object with the changepoint data
}
\examples{
data <- data.frame(X = 1:100, Y = gfpop::dataGenerator(100, c(0.1, 0.3, 0.5, 0.8, 1),
  c(1, 2, 1, 3, 1),
  sigma = 1
))
graph <- gfpop::graph(type = "std", penalty = 15)
changepoint_data <- generate_changepoint(data$Y, graph)
plot_changepoint(data, changepoint_data)
}
