% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_changepoint_plot_helpers.R
\name{plot_changepoint}
\alias{plot_changepoint}
\title{Makes a ggplot object with the changepoint data}
\usage{
plot_changepoint(changepoint_annot)
}
\arguments{
\item{changepoint_annot}{an annotated changepoint dataset}
}
\value{
a ggplot object
}
\description{
Makes a ggplot object with the changepoint data
}
